abstract fn `removeWrappedNative()` to a dedicated file
